[ROCM-SMI] Fix -Werror build failures
- Fix -Werror=write-strings: use char instead of char* for sscanf %c targets
- Fix -Werror=conversion: widen key_position param from uint32_t to size_t
- Fix -Werror=ignored-attributes: use stateless DirCloser functor instead
  of passing closedir directly as a template argument


[ROCM-SMI] Fix -Werror build failures

Add missing #include "rocm_smi/rocm_smi.h" in perf_cntr_read_write.h
to resolve undeclared type errors for rsmi_event_type_t,
rsmi_counter_value_t, and rsmi_event_group_t.

Fix -Wconversion and -Wfloat-conversion warnings across test and
example files by adding explicit static_casts for narrowing
conversions (uint64_t/int64_t to float/double and back).

Remove always-true unsigned >= 0 comparison that triggered
-Wtype-limits in computepartition_read_write.cc.

+++ projects/rocm-smi-lib/include/rocm_smi/rocm_smi_utils.h
@@ -406,8 +406,8 @@ class TagTextContents_t {
   }
 
   decltype(auto) get_structured_data_subkey_by_position(const PrimaryKeyType& prim_key,
-                                                        uint32_t key_position) {
-    auto key_counter = uint32_t(0);
+                                                        size_t key_position) {
+    auto key_counter = size_t(0);
     SecondaryKeyType data_key{};
     if (key_position < (get_structured_subkeys_size(prim_key))) {
       for (const auto& [sec_key, sec_value] : m_structured[prim_key]) {
